Mostly cleaning things, added missing namespaces
