+++ MEP_oM/Equipment/IEquipment.cs
@@ -20,12 +20,6 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using BH.oM.Base;
 
 namespace BH.oM.MEP.Equipment
