+++ MEP_oM/Equipment/Parts/EnergyWheel.cs
@@ -20,13 +20,7 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.ComponentModel;
-
 using BH.oM.Base;
 
 namespace BH.oM.MEP.Equipment.Parts
