+++ MEP_oM/Fragments/GeometryFragment.cs
@@ -20,14 +20,7 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using BH.oM.Base;
-
 using BH.oM.Geometry;
 
 namespace BH.oM.MEP.Fragments
