+++ MEP_oM/System/MaterialFragments/DuctMaterial.cs
@@ -20,20 +20,14 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using BH.oM.Base;
 
-namespace BH.oM.MEP.MaterialFragments
+namespace BH.oM.MEP.System.MaterialFragments
 {
-    public class PipeMaterial : BHoMObject, IMEPMaterial
+    public class DuctMaterial : BHoMObject, IMEPMaterial
     {
-        [Description("Roughness is a measure of the irregularities on the surface of a pipe.")]
+        [Description("Roughness is a measure of the irregularities on the surface of a duct.")]
         public virtual double Roughness { get; set; } = 0;
     }
 }
