+++ MEP_oM/System/MaterialFragments/LiningMaterial.cs
@@ -20,30 +20,20 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.ComponentModel;
-
 using BH.oM.Base;
-using BH.oM.Physical.Materials;
 
-namespace BH.oM.MEP.MaterialFragments
+namespace BH.oM.MEP.System.MaterialFragments
 {
-    [Description("Insulation is the material surrounding a duct, pipe or wire which mitigates the loss of the internal conditions of the fluid within the object.")]
-    public interface IInsulationMaterial : IFragment, IMaterialProperties
+    public class LiningMaterial : BHoMObject, IMEPMaterial, IInsulationMaterial
     {
-        /***************************************************/
-        /**** Properties                                ****/
-        /***************************************************/
-        [Description("RValue is the measure of the resistance of conductive heat loss by the insulation material.")]
-        double RValue { get; set; }
+        [Description("RValue is the measure of the resistance of conductive heat loss by the lining material.")]
+        public virtual double RValue { get; set; } = 0;
 
-        [Description("KValue is the measure of the insulation material's ability to conduct heat (W/m*K), the lower the KValue the better the ability to conduct heat.")]
-        double KValue { get; set; }
+        [Description("KValue is the measure of the lining material's ability to conduct heat (W/m*K), the lower the KValue the better the ability to conduct heat.")]
+        public virtual double KValue { get; set; } = 0;
 
-        /***************************************************/
+        [Description("Roughness is a measure of the irregularities on the surface of the lining.")]
+        public virtual double Roughness { get; set; } = 0;
     }
 }
