+++ MEP_oM/System/MaterialFragments/PipeMaterial.cs
@@ -20,23 +20,14 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using BH.oM.Base;
 
-namespace BH.oM.MEP.MaterialFragments
+namespace BH.oM.MEP.System.MaterialFragments
 {
-    public class InsulationMaterial : BHoMObject, IMEPMaterial, IInsulationMaterial
+    public class PipeMaterial : BHoMObject, IMEPMaterial
     {
-        [Description("RValue is the measure of the resistance of conductive heat loss by the insulation material.")]
-        public virtual double RValue { get; set; } = 0;
-
-        [Description("KValue is the measure of the insulation material's ability to conduct heat (W/m*K), the lower the KValue the better the ability to conduct heat.")]
-        public virtual double KValue { get; set; } = 0;
+        [Description("Roughness is a measure of the irregularities on the surface of a pipe.")]
+        public virtual double Roughness { get; set; } = 0;
     }
 }
