+++ MEP_oM/System/MaterialFragments/WireMaterial.cs
@@ -20,26 +20,14 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using BH.oM.Base;
 
-namespace BH.oM.MEP.MaterialFragments
+namespace BH.oM.MEP.System.MaterialFragments
 {
-    public class LiningMaterial : BHoMObject, IMEPMaterial, IInsulationMaterial
+    public class WireMaterial : BHoMObject, IMEPMaterial
     {
-        [Description("RValue is the measure of the resistance of conductive heat loss by the lining material.")]
-        public virtual double RValue { get; set; } = 0;
-
-        [Description("KValue is the measure of the lining material's ability to conduct heat (W/m*K), the lower the KValue the better the ability to conduct heat.")]
-        public virtual double KValue { get; set; } = 0;
-
-        [Description("Roughness is a measure of the irregularities on the surface of the lining.")]
-        public virtual double Roughness { get; set; } = 0;
+        [Description("The degree to which a specific material conducts electricity, calculated as the ratio of the current density and the material to the electric field that causes the flow of current.")]
+        public virtual double Conductivity { get; set; } = 0;
     }
 }
