+++ MEP_oM/System/MechanicalSystem.cs
@@ -20,18 +20,9 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.ComponentModel;
-
-using BH.oM.MEP.MaterialFragments;
 using BH.oM.MEP.Enums;
 using BH.oM.Base;
-
 
 namespace BH.oM.MEP.System
 {
