+++ MEP_oM/System/SectionProperties/PipeSectionProperty.cs
@@ -20,15 +20,9 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.ComponentModel;
-
 using BH.oM.Base;
-using BH.oM.MEP.MaterialFragments;
+using BH.oM.MEP.System.MaterialFragments;
 
 namespace BH.oM.MEP.System.SectionProperties
 {
