+++ MEP_oM/System/Wire.cs
@@ -20,19 +20,10 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.ComponentModel;
-
 using BH.oM.Base;
-using BH.oM.Analytical.Elements;
-using BH.oM.MEP.System.SectionProperties;
 using BH.oM.Dimensional;
-
 
 namespace BH.oM.MEP.System
 {
